Renamed header files to reflect contents
