+++ src/analyses/variable-sensitivity/full_struct_abstract_object.h
@@ -12,7 +12,7 @@ Author: Thomas Kiley, thomas.kiley@diffblue.com
 #define CPROVER_ANALYSES_VARIABLE_SENSITIVITY_FULL_STRUCT_ABSTRACT_OBJECT_H
 
 #include <analyses/variable-sensitivity/abstract_aggregate_object.h>
-#include <analyses/variable-sensitivity/struct_abstract_object.h>
+#include <analyses/variable-sensitivity/two_value_struct_abstract_object.h>
 #include <iosfwd>
 #include <stack>
 #include <util/sharing_map.h>
